requirements/astro.txt
fastapi
uvicorn[standard]
httpx
SQLAlchemy
aiosqlite
python-dotenv
pydantic
pydantic-settings
kerykeion
pyswisseph
timezonefinder
orjson
